+++ .gitignore
@@ -19,7 +19,6 @@ develop-eggs
 .installed.cfg
 lib
 lib64
-venv
 __pycache__
 
 # Installer logs
@@ -51,6 +50,7 @@ docs.metadata
 
 # Virtual environment
 env/
+venv/
 coverage.xml
 sponge_log.xml
